allow turning off 'cleanup' stage in testMakefakedata.sh via NOCLEANUP environment variable - sometimes useful for testing to leave result-files from testscript Original: 7944f5e39cb66a90f8d88b271b8971bf5ba74a51

+++ lalapps/src/pulsar/Injections/testMakefakedata.sh
@@ -137,6 +137,9 @@ else
 fi
 
 
-## clean up files
-rm -rf $testDIR In.data-test
+## clean up files [allow turning off via 'NOCLEANUP' environment variable
+if [ -z "$NOCLEANUP" ]; then
+    rm -rf $testDIR In.data-test
+fi
+
